requirements.txt
apturl==0.5.2
asgiref==3.7.2
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
CacheControl==0.13.1
cachetools==5.3.2
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
Django==4.2.7
duplicity==0.8.21
fasteners==0.14.1
firebase-admin==6.2.0
future==0.18.2
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.24.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.7
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
proto-plus==1.22.3
protobuf==4.25.1
ptyprocess==0.7.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-firebase @ git+https://github.com/ozgur/python-firebase@0d79d7609844569ea1cec4ac71cb9038e834c355
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
systemd-python==234
typing_extensions==4.8.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
xdg==5
xkit==0.0.0
zipp==1.0.0
